Frontend updates, remove unused assets, restore PR template
