+++ .github/pull_request_template.md
@@ -1,0 +1,16 @@
+## Description
+
+Write a clear documentation of the implemented feature that includes:
+
+- Summary of changes (bug fixes included)
+- Decisions made and reasoning
+- Other relevant discussion
+
+## Checklist
+
+Make sure to tick all the actions below before requesting review
+
+- [ ] Feature is fully done and works
+- [ ] Difficult parts of code have relevant comments
+- [ ] Feature is tested
+- [ ] PR includes clear description for later documentation
