--- frontend/src/constants/index.js
@@ -1,33 +1,0 @@
-import { RiTwitterXLine } from "react-icons/ri";
-import { FaInstagram } from "react-icons/fa";
-import { FaYoutube } from "react-icons/fa6";
-
-export const cards = [
-  {
-    src: "/videos/quiz.mp4",
-    rotation: "rotate-z-[-10deg]",
-    translation: "translate-y-[-5%]",
-  },
-  {
-    src: "/videos/battle.mp4",
-    rotation: "rotate-z-[4deg]",
-  },
-];
-
-export const socials = [
-  {
-    id: 1,
-    icon: RiTwitterXLine,
-    link: "https://twitter.com",
-  },
-  {
-    id: 2,
-    icon: FaInstagram,
-    link: "https://instagram.com",
-  },
-  {
-    id: 3,
-    icon: FaYoutube,
-    link: "https://youtube.com",
-  },
-];
